server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
langchain==0.0.267
langchain-community==0.0.13
pypdf==3.15.1
python-dotenv==1.0.0
faiss-cpu==1.11.0
numpy==1.26.4
pydantic==1.10.8
requests==2.31.0
langdetect==1.0.9
googletrans==3.1.0a0